hiber

+++ src/main/java/lv/javaguru/java2/database/UserDAO.java
@@ -8,9 +8,9 @@ public interface UserDAO {
 
     void create(User user) throws DBException;
 
-    User getById(int id) throws DBException;
+    User getById(Integer id) throws DBException;
 
-    void delete(Long id) throws DBException;
+    void delete(Integer id) throws DBException;
 
     void update(User user) throws DBException;
